+++ BackEnd/MongsilDesk/src/main/java/com/mongdok/desk/model/request/dday/DdayRequest.java
@@ -8,7 +8,7 @@ import lombok.Data;
 
 @Data
 public class DdayRequest {
-	private int deskId;	
+	private long deskId;	
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
 	private Date finishDate;
 	private String title;
