+++ BackEnd/MongsilDesk/src/main/java/com/mongdok/desk/model/request/memo/MemoUpdateRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class MemoUpdateRequest {
-	private int memoId;
+	private long memoId;
 	private String content;
 }
