+++ BackEnd/MongsilDesk/src/main/java/com/mongdok/desk/model/response/dday/DdayResponse.java
@@ -8,7 +8,7 @@ import lombok.Data;
 
 @Data
 public class DdayResponse {
-	private int ddayId;
+	private long ddayId;
 	
 	private String title;
