+++ BackEnd/MongsilDesk/src/main/java/com/mongdok/desk/model/response/guestbook/GuestBookOnlyIdResponse.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class GuestBookOnlyIdResponse {
-	private int boardId;
+	private long boardId;
 }
