+++ BackEnd/MongsilDesk/src/main/java/com/mongdok/desk/model/response/guestbook/GuestBookResponse.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 @Data
 public class GuestBookResponse {
-	private int boardId;
+	private long boardId;
 	
 	private String content;
