+++ BackEnd/MongsilDesk/src/main/java/com/mongdok/desk/service/GuestBookService.java
@@ -8,12 +8,12 @@ import com.mongdok.desk.model.request.guestbook.GuestbookCreateRequest;
 
 public interface GuestBookService {
 
-	ResponseEntity<? extends BasicResponse> getGuestBookById(int boardId);
+	ResponseEntity<? extends BasicResponse> getGuestBookById(long boardId);
 
 	ResponseEntity<? extends BasicResponse> createGuestBook(GuestbookCreateRequest request);
 
 	ResponseEntity<? extends BasicResponse> updateGuestBook(GuestBookUpdateRequest request);
 
-	ResponseEntity<? extends BasicResponse> deleteGuestBook(int boardId);
+	ResponseEntity<? extends BasicResponse> deleteGuestBook(long boardId);
 
 }
